// src/components/parameters/types.ts
import Parameter from "./Parameter";

export type TCurveFunction = (t: number) => number;
export type TParameterRelation = (parameter: Parameter) => number;

export enum eParameterType {
	CONTINUOUS = "continuous",
	DISCRETE = "discrete",
}

export enum eParamAutomationEvent {
	PLAY = "play",
	PAUSE = "pause",
	RESUME = "resume",
	STOP = "stop",
	COMPLETE = "complete",
	UPDATE = "update",
}

export enum eParameterEvent {
	UPDATE = "update",
}

export interface IParameterOptions {
	type: eParameterType;
	name: string;
	range: [number, number];
	value: number;
}

export interface IFadeConfig {
	duration: number;
	curve: TCurveFunction;
}

export interface IValueModulation {
	modulate(value: number, elapsedTime: number): number;
}

export type TTransferFunction = (input: number) => number;

export interface IValueRelation {
	name: string;
	input: number;
	readonly output: number;
}

export interface IValueRelationOptions {
	input: Parameter;
	transferFunction: TParameterRelation;
}

export enum eValueRelationEvent {
	UPDATE = "update",
}

export interface ParamAutomationOptions {
	from: number;
	to: number;
	duration: number;
	curve?: TCurveFunction;
}